refactor(challenge): 최상위 ui/의 단일기능 항목 하강 (구조정리 P5)
NoticeGuideLink는 dashboard 전용이라 challenge/ui/ → dashboard/ui/로 이동.
남은 challenge/ui/(layout, tooltip-question)와 challenge/modal/(LinkChangeConfirmationModal)은
2개+ 기능/페이지가 공유하는 도메인 공용이므로 유지.
→ 최상위 type 폴더는 이제 "진짜 공용"만 보유.
git rename + tsc-주도 import 갱신. 동작 변경 없음.
