--- apps/web/src/domain/challenge/ui/If.tsx
@@ -1,10 +1,0 @@
-import { ReactNode } from 'react';
-
-interface IfProps {
-  condition: boolean;
-  children: ReactNode;
-}
-
-export const If = ({ condition, children }: IfProps) => {
-  return condition ? <>{children}</> : null;
-};
